one test case left

+++ mp_stickers/Image.cpp
@@ -3,7 +3,6 @@
 #include "cs225/PNG.h"
 #include <iostream>
 
-
 // Lighten an Image by increasing the luminance of every pixel by 0.1.
 // This function ensures that the luminance remains in the range [0, 1].    
 void Image::lighten(){
@@ -194,8 +193,8 @@ void Image::scale(unsigned w, unsigned h){
     if (this->width() == 0 || this->height() == 0) {
         this->resize(w,h);
     } else {
-        double ratio1 = w/this->width();
-        double ratio2 = h/this->height();
+        double ratio1 = (double)w/(double)this->width();
+        double ratio2 = (double)h/(double)this->height();
         if (ratio1>ratio2) {
             this->scale(ratio2);
         } else {
